Moving some business logic around
